Update BFAST to use goolf-1.7.20
This updates to using a newer GCC 4.8.4 via goolf-1.7.20 in order to avoid a build
failure for GCC 4.7.2 in goolf-1.4.10.  The build failes  due to a texinfo parsing
bug in GCC documentation files.

This problem appears on systems like CentOS 7 due to stricter parsing in newer
versions of texinfo (5+) supplied by the system environment.

See the bug report for GCC 4.7.2 and patch description

  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=52903