Removed GCC version dependency from texinfo-4.13a easyconfig
Update the easyconfig for texinfo to build using the dummy toolchain.
This removes the version dependency on a particular GCC for the build
and allows a cleaner dependency chain.

The goal of building GCC 4.7.2 with this easyconfig won't depend on
building a later version of GCC first.  It will simply depend on the
compiler already available in the environment.