Make OpenBLAS use the LAPACK version specified in the easyconfig.
The OpenBLAS Makefile only handles large.tgz and timing.tgz and ignores
lapack-x.y.z.tgz. We need to manually unpack the lapack-x.y.z specified
in the easyconfig.

Doing it this way makse it possible to add patches on top of the
unpacked lapack version, which is neccesary to deal with various
compiler problems.