Add an easyconfig for the behave software.
This is a nice little piece of software.
We use it for our vasp test suite, developed by Peter Larsson, NSC LiU

Description of behave:

Behavior-driven development (or BDD) is an agile software development
technique that encourages collaboration between developers, QA and
non-technical or business participants in a software project.